Improved Exception Handling.
Now there is a superclass for DatabaseExceptions. It is an unchecked exception, so it will propagate to the caller method.

Then there are all the specific exceptions that extends that general one.

Fixed the way I handle exceptions for db4o. More informative prints and let the exception be thrown until the caller method (not catching in the implementations).